+++ lib/createImmutableContainer.js
@@ -1,0 +1,1 @@
+// TODO: Implement Higher Order Component that wrapps the immutable logic.
