+++ lib/createImmutableStore.js
