+++ package.json
@@ -2,7 +2,7 @@
   "name": "fluxible-immutable-utils",
   "author": "dhood@yahoo-inc.com",
   "description": "A collection of utilities that provide convenience methods for using Immutable.js with a fluxible based application",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "main": "index.js",
   "repository": {
     "type": "git",
@@ -10,8 +10,8 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "cover": "istanbul cover --dir artifacts -- _mocha tests/unit/ --recursive --reporter spec",
-    "devtest": "mocha tests/unit/ --recursive --reporter spec",
+    "cover": "istanbul cover --dir artifacts -- _mocha tests --recursive --reporter spec",
+    "devtest": "mocha tests --recursive --reporter spec",
     "test": "npm run lint && npm run cover"
   },
   "dependencies": {
